requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-ollama
faiss-cpu
sentence-transformers
pypdf
pymupdf
